New registry style
Instead of storing the registry for each object in the registry
submodule we now instead store them on the classes of the resource type
themselves.  I think this is a bit cleaner, and it lets us ensure that
the resources being added to the registry are of the correct resource
type.