State keeper initial push + a ton of stuff
This is way too big of a checkin.  It started as the implementation of
the state keeper.  Unfortunately I didn't get it in a working state, so
never checked it in, then by the time I got back to it I needed to
standardize on a results format, and that blew up.  In the end it has
all those things, plus the addition of utils/aws_helper which makes
managing the AWS connections a lot easier.

The Keeper currently has an issue where it can get into a race with
reactors - in the end I'll likely include it in the reactor code.