Move to using jinja2 for templating
Our template system needs a few features:

- able to use multi-'dimensional' dictionaries for the mapping
- able to deal with the mapping not having all the variables in the
  template safely

str.format() doesn't have the ability to safely render a template (if a
template calls for something and it doesn't exist, it throws a KeyError
exception)

string.Template can't handle multi-dimensional dictionaries - it can
only take plain variables.

jinja2 meets both of the requirements, and also gives us filters, which
will allow the templates to be even more flexible.  I hated adding a new
dependency, but this was necessary.