Lock API to old Flask & Flask-API
There are some issues with compatibility between newer Flask & the
Flask-API project.  It looks like they are working on them, but
Flask-API also has a link on their README that says its not under active
development right now, so not sure what is going on. This will at least
allow builds to pass, and should provide the existing functionality.